refactor context
