+++ packages/twitter-ui/src/context/TweetContext.tsx
@@ -2,13 +2,15 @@ import { createContext, useContext, useState } from "react";
 import { ITweet } from "../types/tweet";
 import { tweetsData } from "../tweetsData";
 
-export const TweetContext = createContext<TweetContextValues | undefined>(
-  undefined
-);
+export const TweetContext =
+  createContext<TweetContextValues | undefined>(undefined);
 
 interface TweetContextValues {
   tweets: ITweet[];
-  setTweets: React.Dispatch<React.SetStateAction<ITweet[]>>;
+  addTweet: (tweet: ITweet) => void;
+  addRetweet: (tweet: ITweet) => void;
+  incrementFavorite: (tweetId: number) => void;
+  decrementFavorite: (tweetId: number) => void;
 }
 
 interface TweetContextProviderChildren {
@@ -19,13 +21,53 @@ export const TweetContextProvider: React.FC<TweetContextProviderChildren> = ({
   children,
 }) => {
   const [tweets, setTweets] = useState<ITweet[]>(tweetsData);
-  const value: TweetContextValues = { tweets, setTweets };
+
+  const incrementFavorite = (id: number) => {
+    const findAndIncrementTweetById = tweets.map((tweet) => {
+      return tweet.id === id
+        ? {
+            ...tweet,
+            favoriteCount: tweet.favoriteCount + 1,
+          }
+        : tweet;
+    });
+    setTweets(findAndIncrementTweetById);
+  };
+
+  const decrementFavorite = (id: number) => {
+    const findAndDecrementTweetById = tweets.map((tweet) => {
+      return tweet.id === id
+        ? {
+            ...tweet,
+            favoriteCount: tweet.favoriteCount - 1,
+          }
+        : tweet;
+    });
+
+    setTweets(findAndDecrementTweetById);
+  };
+
+  const addTweet = (tweet: ITweet) => {
+    setTweets((prevTweets) => [tweet, ...prevTweets]);
+  };
+
+  const addRetweet = (tweet: ITweet) => {
+    setTweets([{ ...tweet, id: tweet.id, type: "retweet" }, ...tweets]);
+  };
+
+  const value: TweetContextValues = {
+    tweets,
+    addTweet,
+    addRetweet,
+    incrementFavorite,
+    decrementFavorite,
+  };
   return (
     <TweetContext.Provider value={value}>{children}</TweetContext.Provider>
   );
 };
 
-export function useTweetContext() {
+export function useTweets() {
   const context = useContext(TweetContext);
   if (context === undefined) {
     throw new Error(
